refactor(JOE-942): extract supervisor + project-binding helpers
Move roadmap supervisor and project-binding/completion helpers into leaf
modules under work-store/ so the façade stays acyclic. Ratchet work-store
file LOC budget 4900→4500 (~4424 lines after extract).
